hushh-adk-agents/pyproject.toml
[project]
name = "hushh-kai-demo"
version = "0.1.0"
description = "Hushh Kai Deal Optimizer - MVP Demo using ADK + MCP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk==0.3.3",
    "fastmcp==2.11.3",
    "google-adk==1.13.0",
    "google-genai>=1.17.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
]
